Only apply safe_erb to ActionView::Renderable
We were applying safe_erb to all ERB invocations that weren't on our
whitelisted set of file extensions.  But this broke script/generate.
So after a bit of tweaking, this patch limits safe_erb to all instances
of ActionView::Renderable that aren't on our whitelist.

I've done a fair bit of testing, using both automated tests and manual
tests.  So far, safe_erb seems to be working.

This code is fairly fragile, because the internals of ActionView are
pretty weird, especially when view caching is turned on.  It's fairly
hard to write a unit test which proves that safe_erb is actually
working when config.action_controller.perform_caching is true.